+++ src/containers/MoreViewPage/components/Hr.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import styled from 'styled-components';
+
+const Hr = () => {
+  return (
+    <div>
+      <Horizon />
+    </div>
+  );
+};
+
+export default Hr;
+
+const Horizon = styled.hr`
+  margin-bottom: 20px;
+`;
